producerservice/requirements.txt
kafka-python
ijson
